backend/requirements.txt
fastapi==0.123.0
uvicorn==0.38.0
sqlalchemy==2.0.44
pydantic==2.12.5
python-multipart==0.0.20
python-jose==3.5.0
passlib==1.7.4

google-generativeai==0.8.3
python-dotenv==1.2.1
chromadb==1.3.5
